pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = [
    "hatchling",
]

[project]
name = "allianceauth-afat"
description = "Another Fleet Activity Tracking tool for Alliance Auth"
readme = "README.md"
keywords = [
    "allianceauth",
    "eveonline",
    "template",
    "theme",
]
license = { file = "LICENSE" }
authors = [
    { name = "Peter Pfeufer", email = "develop@ppfeufer.de" },
]
requires-python = ">=3.10"
classifiers = [
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 4.2",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
]
dynamic = [
    "version",
]
dependencies = [
    "allianceauth>=4.6,<5",
    "allianceauth-app-utils>=1.14.1",
    "unidecode>=1.3.4",
]
optional-dependencies.tests-allianceauth-latest = [
    "allianceauth-securegroups",
    "coverage",
    "django-webtest",
    "fittings",
]
urls.Changelog = "https://github.com/ppfeufer/allianceauth-afat/blob/master/CHANGELOG.md"
urls.Documentation = "https://github.com/ppfeufer/allianceauth-afat/blob/master/README.md"
urls.Donations = "https://ko-fi.com/ppfeufer"
urls.Homepage = "https://github.com/ppfeufer/allianceauth-afat"
urls.Source = "https://github.com/ppfeufer/allianceauth-afat.git"
urls.Tracker = "https://github.com/ppfeufer/allianceauth-afat/issues"

[tool.hatch.build]
include = [
    "/afat",
]

[tool.hatch.version]
path = "afat/__init__.py"
